move rowcomp to components
