+++ components/RowComp.tsx
@@ -1,0 +1,19 @@
+import { View } from "react-native";
+
+type RowCompProps = {
+  children: React.ReactNode;
+};
+
+export const ViewRow = ({ children }: RowCompProps) => {
+  return (
+    <View
+      style={{
+        flexDirection: "row",
+        alignItems: "center",
+        justifyContent: "space-between",
+      }}
+    >
+      {children}
+    </View>
+  );
+};
